% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmetics.R
\name{create_tip_names}
\alias{create_tip_names}
\title{teste}
\usage{
create_tip_names(demes_matrix, origin)
}
\arguments{
\item{region_n}{teste}

\item{exog_n}{teste}
}
\value{
tesre
}
\description{
teste
}
\examples{
teste
}
