% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlformatting.R
\name{format_xml_beast1}
\alias{format_xml_beast1}
\title{Generates a runnable XML file to run discrete phylogeography in beast1}
\usage{
format_xml_beast1(xmlfn, fastafn)
}
\arguments{
\item{xmlfn}{The file name of the XML skeleton}

\item{fastafn}{The file name of the sequence data in FASTA format}
}
\value{
Character string of runnable XML individual XMLs for each sequence
   alignment will be written to disk
}
\description{
This function will insert the sequence data, tip dates and
location to a XML skeleton
}
\details{
The names of the sequence data and trees must match and must be in the
standard form for inferring tip dates. The XML skeleton must have these
entries to be overwritten:
SEQUENCES
TRAITS

A XML file will be generated for each sequence alignment provided
}
