% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlformatting.R
\name{format_xml_phydyn}
\alias{format_xml_phydyn}
\title{Generates a runnable XML file to run phydyn in beast2}
\usage{
format_xml_phydyn(xmlfn, fastafn)
}
\arguments{
\item{xmlfn}{The file name of the XML skeleton}

\item{fastafn}{The file name of the sequence data in FASTA format}
}
\value{
Character string of runnable XML Individual XMLs for each starting
   tree will be written to disk
}
\description{
This function will insert the sequence data, tip dates to a XML skeleton
}
\details{
The names of the sequence data and trees must match and must be in the
standard form for inferring tip dates. The XML skeleton must have these
entries to be overwritten:
SEQUENCES
DATE_TRAIT

A XML file will be generated for each sequence alignment provided
}
