% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_states}
\alias{sample_states}
\title{Random sample states}
\usage{
sample_states(n_compartments, n_size, compartment_names)
}
\arguments{
\item{n_compartments}{Integer for the total number of compartments.}

\item{n_size}{Integer for the total number of individuals to random sample
their states.}

\item{compartment_names}{Vector with the string names for each deme or
compartment.}
}
\value{
Matrix with 0 and 1s. Value of 1 correspond to the sampled
   deme for each individual. Each colunm in the matrix is a deme in the model;
   and each row correspond to an individual.
}
\description{
This function will take the total number of compartments (or demes) in the
model and random select the number of diagnosed individuals not on ART.
}
\examples{
ss <- sample_states(3, 10, c("care1", "care2", "care3"))
}
