% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/migr.functions.R
\name{test5}
\alias{test5}
\title{Histogram of salaries by hours per week}
\usage{
test5(x, main = "Salaries by hours per week", cex = 0.4)
}
\description{
Histogram of salaries by hours per week
}

