% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/migr.functions.R
\name{test6}
\alias{test6}
\title{Histogram of salaries by migration status}
\usage{
test6(x, main = "Salaries by migration status", cex = 0.4)
}
\description{
Histogram of salaries by migration status
}

