% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/migr.functions.R
\name{test7}
\alias{test7}
\title{Histogram of salaries by migration status}
\usage{
test7(x, what = "", main = paste("Salaries by", what), cex = 0.4)
}
\description{
Histogram of salaries by migration status
}

